requirements.txt
numpy==1.23.5
opencv-python==4.8.1.78
torch==2.0.1
torchvision==0.15.2
onnx==1.14.0
onnxruntime-gpu==1.16.0
stable-baselines3==2.1.0
gymnasium==0.29.1
pyyaml==6.0.1
tensorboard==2.14.0
matplotlib==3.7.2
pandas==2.0.3
pillow==10.0.1
ultralytics==8.0.200
tqdm==4.66.1
psutil==5.9.5